+++ internal/handler/transaction.go
@@ -1,109 +1,43 @@
 package handler
 
 import (
-	"database/sql"
-	"log"
-	"os"
+	"net/http"
+	"unify/internal/database"
 	"unify/internal/models"
+
+	"github.com/gin-gonic/gin"
 )
 
-func PostTransaction(req models.TransactionRequestPayload) (res models.Transaction) {
-	// データベースのハンドルを取得する
-	db, err := sql.Open("mysql", os.Getenv("MYSQL_PASS")+":"+os.Getenv("MYSQL_USER")+"@tcp(localhost:3306)/go_test")
-
-	if err != nil {
-		// ここではエラーを返さない
-		log.Fatal(err)
+func PostTransaction(c *gin.Context) {
+	var newCustomer models.TransactionRequestPayload
+	if err := c.BindJSON(&newCustomer); err != nil {
+		return
 	}
-	defer db.Close()
-	TransactionDate := GetDate()
-	TransactionId := GettransactionId()
-	// SQLの準備
-	ins, err := db.Prepare("INSERT INTO transaction VALUES(?,?,?,?,?,?)")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer ins.Close()
-
-	// SQLの実行
-	_, err = ins.Exec(req.UID, TransactionId, req.ItemId, TransactionDate, req.Count, req.IsFinished)
-	if err != nil {
-		log.Fatal(err)
-	}
-	return GetTransaction(req.UID)
+	res := database.PostTransaction(newCustomer)
+	c.IndentedJSON(http.StatusOK, res)
 }
 
-func GetTransaction(uid string) (res models.Transaction) {
-	// データベースのハンドルを取得する
-	db, err := sql.Open("mysql", os.Getenv("MYSQL_PASS")+":"+os.Getenv("MYSQL_USER")+"@tcp(localhost:3306)/go_test")
-
-	if err != nil {
-		// ここではエラーを返さない
-		log.Fatal(err)
+func UpdateTransaction(c *gin.Context) {
+	h := new(database.PatchRequestPayload)
+	if err := c.BindJSON(&h); err != nil {
+		return
 	}
-	defer db.Close()
-
-	// SQLの実行
-	rows, err := db.Query("SELECT * FROM transaction WHERE uid = ?", uid)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer rows.Close()
-	var Customer models.Transaction
-	// SQLの実行
-	for rows.Next() {
-		err := rows.Scan(&Customer.UID, &Customer.ItemId, &Customer.TransactionDate, &Customer.Count, &Customer.IsFinished)
-		if err != nil {
-			panic(err.Error())
-		}
-	}
-	return Customer
+	h.Patch("transaction", false, "transactionid")
+	c.JSON(http.StatusOK, database.GetItemList())
 }
-func UpdateTransaction(req models.Transaction) (res models.Transaction) {
-	// データベースのハンドルを取得する
-	db, err := sql.Open("mysql", os.Getenv("MYSQL_PASS")+":"+os.Getenv("MYSQL_USER")+"@tcp(localhost:3306)/go_test")
 
-	if err != nil {
-		// ここではエラーを返さない
-		log.Fatal(err)
+func GetTransaction(c *gin.Context) {
+	uid := c.Query("uid")
+	var response = database.GetTransaction(uid)
+	if response.UID == "" {
+		c.IndentedJSON(http.StatusNotFound, gin.H{"message": "customer not found"})
+		return
 	}
-	defer db.Close()
-
-	// SQLの準備
-	upd, err := db.Prepare("UPDATE transaction SET id = ?, transactionDate = ?, count = ?, isFinished = ? WHERE uid = ?")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer upd.Close()
-	// SQLの実行
-	_, err = upd.Exec(req.ItemId, req.TransactionDate, req.Count, req.IsFinished, req.UID)
-	if err != nil {
-		log.Fatal(err)
-	}
-	return GetTransaction(req.UID)
+	c.IndentedJSON(http.StatusOK, response)
 }
-func DeleteTransaction(uid string) (s string) {
-	// データベースのハンドルを取得する
-	db, err := sql.Open("mysql", os.Getenv("MYSQL_PASS")+":"+os.Getenv("MYSQL_USER")+"@tcp(localhost:3306)/go_test")
 
-	if err != nil {
-		// ここではエラーを返さない
-		log.Fatal(err)
-	}
-	defer db.Close()
-
-	// SQLの実行
-	del, err := db.Prepare("DELETE FROM transaction WHERE uid = ?")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer del.Close()
-
-	// SQLの実行
-	_, err = del.Exec(uid)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	return
+func DeleteTransaction(c *gin.Context) {
+	transactionid := c.Query("transactionid")
+	database.Delete("transaction", "ID", transactionid)
+	c.JSON(http.StatusOK, gin.H{"message": "transaction was successfully deleted"})
 }
