+++ internal/models/item.go
@@ -1,0 +1,12 @@
+package models
+
+type Item struct {
+	ID          string `json:"id"`
+	Price       int    `json:"price"`
+	Name        string `json:"Name"`
+	Stonesize   int    `json:"Stonesize"`
+	Minlength   int    `json:"Minlength"`
+	Maxlength   int    `json:"Maxlength"`
+	Decsription string `json:"Description"`
+	Keyword     string `json:"Keyword"`
+}
